refactor ReviewScoreIndicator
